+++ shell_scripts/de/requirements.txt
@@ -1,4 +1,3 @@
 germalemma
 django-environ
 pandas
-treetaggerwrapper
