+++ shell_scripts/en/requirements.txt
@@ -1,4 +1,3 @@
 nltk
-pyknp
 pandas
 django-environ
